requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pyhunter==1.7
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.0
urllib3==1.25.8
